refactor-shosesearch-jobeomjun---이미지 검색 리팩토링3차 완료
