+++ src/components/choose-shose/image-search/analyzeimage/succesproduct/succesproduct.css.ts
@@ -83,55 +83,3 @@ export const Product_ProductName = style({
   letterSpacing: '-0.003em',
   color: theme.color.black,
 });
-export const Product_ProductPrice = style({
-  fontWeight: '600',
-  fontSize: '16px',
-  lineHeight: '20px',
-  letterSpacing: '-0.003em',
-  color: theme.color.black,
-});
-
-export const Product_SimilarProductContainer = style({
-  width: '100%',
-  height: '100%',
-  paddingTop: '29px',
-  paddingBottom: '108px',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
-
-export const Product_SimilarProductButton = style({
-  height: '13%',
-  position: 'fixed',
-  bottom: '0',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
-
-export const Product_ScrollableContent = style({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  width: '100%',
-  height: '100%',
-  overflowY: 'scroll',
-
-  selectors: {
-    '&::-webkit-scrollbar': {
-      width: '6px', // 스크롤바의 너비를 줄임
-    },
-    '&::-webkit-scrollbar-thumb': {
-      backgroundColor: 'rgba(0, 0, 0, 0.4)', // 스크롤바의 색상
-      borderRadius: '10px', // 스크롤바의 모서리를 둥글게
-    },
-    '&::-webkit-scrollbar-track': {
-      backgroundColor: 'transparent', // 스크롤바 트랙의 배경
-    },
-  },
-});
-
-export const Product_ItemCardsContainer = style({
-  width: '90%',
-});
