+++ src/stores/useImageSearchStore.ts
@@ -1,9 +1,5 @@
 import { create } from 'zustand';
 
-type ImageSearchStore = {
-  handleClickAgain: () => void;
-};
-
 type ImageSearchStateStore = {
   isAnalyze: boolean;
   isSuccess: boolean;
@@ -11,19 +7,12 @@ type ImageSearchStateStore = {
   setIsState: (changedState: Partial<ImageSearchStateStore>) => void;
 };
 
-const useImageSearchStore = create<ImageSearchStore & ImageSearchStateStore>(set => ({
+const useImageSearchStore = create<ImageSearchStateStore>(set => ({
   isAnalyze: false,
   isSuccess: false,
   isSimilar: false,
   isGallery: false,
   setIsState: changedState => set(state => ({ ...state, ...changedState })),
-
-  handleClickAgain: () => {
-    set(state => {
-      state.setIsState({ isAnalyze: false, isSuccess: false });
-      return {};
-    });
-  },
 }));
 
 export default useImageSearchStore;
